Extracted top bar in preparation for the refactor
